+++ js/upload-image.js
@@ -3,6 +3,7 @@ import {createValidation} from './upload-validation.js';
 import {addEffectsSetting} from './upload-effects.js';
 import {addScalingController} from './upload-scale-image.js';
 import {pristineDefaultConfig, SubmitButtonText} from './constats.js';
+import {showError} from './upload-error.js';
 import {sendData} from './load-data.js';
 
 const uploadForm = document.querySelector('.img-upload__form');
@@ -24,9 +25,6 @@ const uploadWrapper = uploadForm.querySelector('.img-upload__wrapper');
 const successTemplate = document.querySelector('#success')
   .content
   .querySelector('.success');
-const errorTemplate = document.querySelector('#error')
-  .content
-  .querySelector('.error');
 
 addScalingController(scaleControlInput, imageToUpload, increaseScaleButton, decreaseScaleButton);
 
@@ -80,46 +78,6 @@ const showSuccess = () => {
   successButton.addEventListener('click', removeSuccessMessage);
 
   uploadOverlay.appendChild(successElement);
-};
-
-const showError = (message) => {
-  const errorElement = errorTemplate.cloneNode(true);
-  const errorInnerContainer = errorElement.querySelector('.error__inner');
-  const errorTitle = errorElement.querySelector('.error__title');
-  const errorButton = errorElement.querySelector('.error__button');
-  errorTitle.textContent = message;
-  uploadWrapper.classList.add('hidden');
-  document.removeEventListener('keydown', cancelUploadByKeydown);
-  let returnToFormOnEscapeClick = () => {};
-
-  const removeErrorMessage = () => {
-    errorButton.removeEventListener('click', removeErrorMessage);
-    document.removeEventListener('keydown', returnToFormOnEscapeClick);
-    document.addEventListener('keydown', cancelUploadByKeydown);
-    uploadWrapper.classList.remove('hidden');
-    errorElement.remove();
-  };
-
-  returnToFormOnEscapeClick = (evt) => {
-    if (isEscapeKey(evt)) {
-      removeErrorMessage();
-    }
-  };
-
-  document.addEventListener('keydown', returnToFormOnEscapeClick);
-
-  const onOutsideErrorContainerClick = (evt) => {
-    const outsideErrorContainerClick = evt.composedPath().includes(errorInnerContainer) === false;
-    if (outsideErrorContainerClick) {
-      removeErrorMessage();
-      document.removeEventListener('click', onOutsideErrorContainerClick);
-    }
-  };
-
-  document.addEventListener('click', onOutsideErrorContainerClick);
-  errorButton.addEventListener('click', removeErrorMessage);
-
-  uploadOverlay.appendChild(errorElement);
 };
 
 noUiSlider.create(sliderControlContainer, {
@@ -183,4 +141,4 @@ const initUploadImageForm = () => {
   });
 };
 
-export {initUploadImageForm};
+export {initUploadImageForm, uploadWrapper, cancelUploadByKeydown, uploadOverlay};
